+++ cli/Cargo.toml
@@ -1,12 +1,3 @@
-[package]
-name = "cli"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-csv = "1.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-clap = { version = "4.0", features = ["derive"] }
-anyhow = "1.0"
-thiserror = "1.0"
+[workspace]
+resolver = "3"
+members = ["cli", "motorsport"]
