+++ cli/motorsport/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "motorsport"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
